Extract shared tmux test support
The tmux discovery suite cannot be split by responsibility while its reusable fixtures remain file-private. Move the shared builders, surface stubs, queues, and launch helper behind an explicit test-only boundary without changing coverage or production access.

This keeps the later discovery, presentation, and project-removal split mechanical and reviewable.
